requirements.txt
streamlit
praw
plotly
nltk
vaderSentiment
requests
python-dotenv
